+++ src/NovelRT.Java/core/src/main/kotlin/com/github/novelrt/ecs/Types.kt
@@ -1,7 +1,7 @@
 package com.github.novelrt.ecs
 
-typealias EntityId = UInt
-typealias ComponentTypeId = UInt
-typealias PoolId = UInt
-typealias SparseSetKey = UInt
+typealias EntityId = ULong
+typealias ComponentTypeId = ULong
+typealias PoolId = ULong
+typealias SparseSetKey = ULong
 // TODO: Fix FumoCement so these can be ULong
